use safe int conversion

+++ logstash-core/src/main/java/org/logstash/instrument/metrics/timer/Util.java
@@ -10,6 +10,6 @@ public class Util {
     }
 
     static int subMilliExcessNanos(final long excessNanos) {
-        return (int)Math.floorMod(excessNanos, NANOS_PER_MILLI);
+        return Math.toIntExact(Math.floorMod(excessNanos, NANOS_PER_MILLI));
     }
 }
